Move triangulate to own module, return indices instead of mesh
